[ref] Separate tray menu class
