+++ cozyfuse/interface/app.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: UTF-8 -*-
 #
-# generated by wxGlade 0.6.8 on Thu Aug  7 12:49:06 2014
+# generated by wxGlade 0.6.8 on Thu Aug  7 13:00:55 2014
 #
 
 # This is an automatically generated file.
