+++ cozyfuse/interface/cozy_client_template.wxg
@@ -1,9 +1,8 @@
 <?xml version="1.0"?>
-<!-- generated by wxGlade 0.6.8 on Thu Aug  7 12:49:04 2014 -->
+<!-- generated by wxGlade 0.6.8 on Thu Aug  7 13:00:53 2014 -->
 
 <application path="/home/kload/dev/cozy-fuse-linux/cozyfuse/interface/" name="" class="" option="1" language="python" top_window="cozy_frame" encoding="UTF-8" use_gettext="1" overwrite="1" use_new_namespace="1" for_version="2.8" is_template="0" indent_amount="4" indent_symbol="space" source_extension=".cpp" header_extension=".h">
     <object class="CozyFrame" name="cozy_frame" base="EditDialog">
-        <extracode>TRAY_TOOLTIP = 'System Tray Demo'\nTRAY_ICON = 'icon/icon.png'\nSMALL_TRAY_ICON = 'icon/small_icon.png'\n\n\ndef create_menu_item(menu, label, func):\n    item = wx.MenuItem(menu, -1, label)\n    menu.Bind(wx.EVT_MENU, func, id=item.GetId())\n    menu.AppendItem(item)\n    return item\n\n\nclass TaskBarIcon(wx.TaskBarIcon):\n    def __init__(self):\n        super(TaskBarIcon, self).__init__()\n        self.set_icon(TRAY_ICON)\n        self.Bind(wx.EVT_TASKBAR_LEFT_DOWN, self.on_left_down)\n\n    def CreatePopupMenu(self):\n        menu = wx.Menu()\n        create_menu_item(menu, 'Say Hello', self.on_hello)\n        menu.AppendSeparator()\n        create_menu_item(menu, 'Exit', self.on_exit)\n        return menu\n\n    def set_icon(self, path):\n        icon = wx.IconFromBitmap(wx.Bitmap(path))\n        self.SetIcon(icon, TRAY_TOOLTIP)\n\n    def on_left_down(self, event):\n        print 'Tray icon was left-clicked.'\n\n    def on_hello(self, event):\n        import debug\n        self.set_icon(SMALL_TRAY_ICON)\n        print 'Hello, world!'\n\n    def on_exit(self, event):\n        wx.CallAfter(self.Destroy)</extracode>
         <style>wxDEFAULT_DIALOG_STYLE</style>
         <icon>icon/icon.png</icon>
         <title>Cozy client configuration</title>
